requirements.txt
pymongo>=4.6.0
redis>=5.0.0
Faker>=24.0.0
python-dotenv>=1.0.0
streamlit
redis
plotly
pandas